Simplified requirements and renamed the package to what it should do.
