+++ src/main/java/com/rolfje/anonimatron/file/RecordReader.java
@@ -1,0 +1,9 @@
+package com.rolfje.anonimatron.file;
+
+/**
+ * Provides records to be anonymized the {@link FileAnonymizerService}.
+ */
+public interface RecordReader {
+	boolean hasRecords();
+	Record read();
+}
