+++ src/main/java/com/rolfje/anonimatron/file/RecordWriter.java
@@ -1,0 +1,8 @@
+package com.rolfje.anonimatron.file;
+
+/**
+ * Consumes records which were anonymized by {@link FileAnonymizerService}.
+ */
+public interface RecordWriter {
+	void write(Record record);
+}
