--- src/main/java/com/rolfje/anonimatron/stream/RecordReader.java
@@ -1,9 +1,0 @@
-package com.rolfje.anonimatron.stream;
-
-/**
- * Provides records to be anonymized the {@link StreamAnonymizerService}.
- */
-public interface RecordReader {
-	boolean hasRecords();
-	Record read();
-}
