--- src/main/java/com/rolfje/anonimatron/stream/RecordWriter.java
@@ -1,8 +1,0 @@
-package com.rolfje.anonimatron.stream;
-
-/**
- * Consumes records which were anonymized by {@link StreamAnonymizerService}.
- */
-public interface RecordWriter {
-	void write(Record record);
-}
